Moved pages into /pages
